Scrap bpm reference
The current "conversion" to bpm doesn't work since it entirely depends on how
fast the browser can get through one frame refresh cycle, so scrap it for now.
The fix would be to regulate the framerate of the app (i.e., the rate at which
the planet (etc.) positions are updated onscreen).